+++ packages/kiapi-proxy/src/kiapi_proxy/core/config/_helpers/get_user_settings_path.py
@@ -1,7 +1,0 @@
-from pathlib import Path
-
-from kiarina.utils.app import user_directory
-
-
-def get_user_settings_path() -> Path:
-    return user_directory.get_user_config_dir() / "settings.yaml"