+++ packages/kiapi-relay/src/kiapi_relay/core/_types/relay_method.py
@@ -1,3 +1,0 @@
-from typing import Literal
-
-type RelayMethod = Literal["DELETE", "GET", "PATCH", "POST", "PUT"]